Changed filename, changed display (message indicating that the user does not have any households, and guide them to either create a new household or join an existing one)
